Pre-process out physics dimensions, pools, and options without -DDO_PHYSICS
The inclusion of physics dimensions, pools, and namelist options in Registry
files now depends on the definition of DO_PHYSICS being present in
the pre-processing flags.

Along with the exclusion of physics dimensions, pools, and options when
DO_PHYSICS is not defined, references to the diag_physics and tend_physics
pools, as well as physics namelist options, are now pre-processed out
of the code.

Some additional changes are needed to properly handle 'rthdynten' in
the dynamics, since this field comes from the 'tend_physics' pool, which may
not exist.

Also, the PV diagnostics module now excludes code and fields for PV budgets
when DO_PHYSICS is not defined, since these depend on fields from physics.

** NB: In order to correctly run with no physics, there are additional
   changes that are needed beyond those in this commit. Specifically,
   changes like those in 8cb5065 (PR #193) are needed.