Allow mpas_bootstrap_framework_phase1 to use an externally provided PIO file_desc_t
When MPAS is a component in a larger system, it may be necessary to "bootstrap"
the MPAS infrastructure from an externally provided file descriptor. This commit
adds an optional PIO file_desc_t argument to mpas_bootstrap_framework_phase1, which
is then passed to MPAS_io_open. The MPAS_io_open routine now uses the optional
PIO file_desc_t if provided, and opens the file indicated by the filename argument
as before if no PIO file_desc_t argument is provided.