Don't call PIO_closefile in MPAS_io_close for externally provided PIO filedesc
In order to avoid closing a PIO file_desc_t that was not opened by MPAS_io_open
(because this external file_desc_t was provided as an optional argument in
the call to MPAS_io_open), this commit adds a new member, external_file_desc,
to the MPAS_IO_Handle_type type. This member is set to .true. if an external
PIO file_desc_t was provided, and if .true. in the call to MPAS_io_close,
PIO_closefile is not called.