Add pools for CAM dynamics import and export states
The fields defined in the new dyn_in and dyn_out pools may be directly
referenced by CAM-MPAS's dynamics import and export states, since
the members of those states are pointers.

Eventually, redundant versions of fields (e.g, uReconstructZonal and ux) that
already exist elsewhere in the Registry.xml file should be avoided, but
the final organization of CAM-MPAS's import and export states is yet unknown.