Remove superfluous if-test around call to atm_advance_scalars_work
The atm_advance_scalars routine previously had a test on the value of
local_advance_density to determine how to call the atm_advance_scalars_work.
However, both calls to atm_advance_scalars_work are identical after
the changes in the preceding commit, and so the if-test can be eliminated.