Enable CAM to override constants when MPAS-A is compiled as a CAM dycore
In order to allow MPAS-Atmosphere to use identical values for constants to CAM
when it is compiled for use as a CAM dycore, the definitions of parameters in
the mpas_constants module now depend on whether the preprocessing macro
MPAS_CAM_DYCORE is defined at build time. If this macro is not defined,
the parameters in mpas_constants are defined as before; otherwise,
the definitions in the mpas_constants module come from CAM's physconst module,
in which case the "constants" are actually variables (CAM allows for runtime
modification of constants).

Additionally, to provide derived constants that are used in the MPAS-A dycore
but not defined in CAM's physconst module, the dummy routine in
the mpas_constants module has been replaced with a routine,
mpas_constants_compute_derived, that can be called to compute constants that
are derived from others after CAM defines constants in its physconst module
at runtime.