Support runtime definition of scalars when MPAS-A is a CAM dycore
This commit makes changes to enable CAM to define the set of scalar constituents
for the MPAS-A dycore completely at runtime.

When MPAS-A is compiled as a CAM dycore, the macro MPAS_CAM_DYCORE is defined,
in which case the scalars and scalars_tend var_arrays will be pre-processed out
of the Registry.xml file so that no Registry-defined scalars will exist.

Further, CAM will add a config, 'cam_pcnst', which will cause the atm_setup_block
routine to call a new routine, atm_allocate_scalars, in the atm_core_interface
module. The atm_allocate_scalars routine will allocate the scalars and
scalars_tend var_arrays in the state and tend pools, respectively, but the actual
scalar constituents to appear in specific indices of these var_arrays is defined
only later by the CAM-MPAS interface layer.