New initialization case to generate a 3-d grid for use with CAM-MPAS
This new initialization case takes a unit-sphere input, a specification of
the zeta levels, and the surface geopotential field, and produces a 3-d grid
suitable for use with CAM-MPAS.

The zeta levels must be provided in a text file, which is identified by
the config_specified_zeta_levels option in the &vertical_grid namelist group.

The surface geopotential must be provided in a netCDF file described in
an input stream, 'cam_topo', where the PHIS field is dimensioned by nCells.

The new case is numbered 13, since there is other work in progress that
uses case numbers 10, 11, and 12.