Add namelist and streams filename parameters to mpas_init
This alters the signature of the mpas_init subroutine in mpas_subdriver.F,
adding two optional parameters that allow the caller to pass in filenames for
the namelist and streams file. This allows the caller to specify different files
to be used, other than namelist.atmosphere and streams.atmosphere in the working
directory. This functionality is critical for several use-cases in MPAS-JEDI,
including the ConvertState application, which interpolates data between two MPAS
meshes of different resolutions.