Fix indentation of else-if statement on line 110 of mpas_timekeeping.F
The else-if statement on line 110 of mpas_timekeeping.F now has the same
indentation as the other if/else-if/else statements in its code block.

Note: This fix-up is being made as a separate commit rather than as an amendment
to commit 894c68f, since 894c68f has already been merged to another branch and
it would be nice to avoid two commits in the repository with identical content
but different hashes.