Fix negative time interval strings in mpas_get_timeInterval with external ESMF
The mpas_get_timeInterval routine didn't correctly generate time interval
strings for negative time intervals when using the external ESMF library.
This commit adds checks on the components (days, hours, minutes, and seconds)
of the time interval to determine whether the time interval string needs to
be prepended with a minus sign.

This commit also employs a suggestion from @philipwjones to use a single format
statement when constructing the time interval string, rather than switching
between format statements depending on the magnitude of the seconds component.