Enable dynamics-only MPAS-Atmosphere build by simply undefining PHYSICS variable
Although previous commits enabled the MPAS-Atmosphere model to be compiled in
a dynamics-only configuration provided that manual changes were made to
Makefiles, these changes to the Makefiles were neither obvious nor documented.

With this commit, building a dynamics-only model can be accomplished by simply
deleting or commenting-out the definition of the PHYSICS variable near the top
of the src/core_atmosphere/Makefile file. Make dependencies are then set
appropriately based on the definition status of PHYSICS to provide a correct
full-physics or dynamics-only build.