Remove unnecessary if-test around cam_pcnst logic in atm_setup_block
The block of code to test on the presence of a configuration option named
'cam_pcnst' and to call atm_allocate_scalars if that option existed was
previously enclosed in an if-test on the association status of the configs pool.
This if-test is unnecessary, as both CAM-MPAS and stand-alone MPAS-Atmosphere
always allocate the MPAS domain object and set-up the configs pool before
the atm_setup_block routine is called.