Convert halo exchanges in MPAS-A to use group exchange routines from mpas_dmpar
Previously, MPAS-A used older halo exchange routines that handled a single field
per call. With this commit, halos are now exchanged using newer (c. 2016)
routines in the mpas_dmpar module that can communicate halos for a group of
fields in a single call.

As part of the changes in this commit, the use of temporary, ad hoc local fields
for exchanging halos of fields in the monotonic scalar transport has been
reworked to make use of scratch arrays.

In cases where only a single field needs to have its halo exchanged at a
specific point in the code, that halo exchange has still been updated to use the
group halo exchange routines by creating an exchange group with a single member.

Note: In this commit, the halo exchanges in diagnostics (specifically, isobaric
diagnostics and PV diagnostics) have not been modified.