Remove mgr dummy argument from abstract subroutine
This commit removes the `mgr` dummy argument from the abstract subroutines that
were introduced in 82e529b5. The mgr dummy argument is being removed as a
workaround was found to avoid creating warnings when the mgr argument was not
needed for specific fields. Fields that will be needed in the mgr pool will now
be passed in as module variables.