Fix bug in vertical interp of r.h. and s.h. when levels are given top-to-bottom
The code in the real-data initialization routine for vertically interpolating
relative humidity and specific humidity assumed that first-guess levels would
be given in bottom-to-top order when attempting to vertically extrapolate to
model levels below the lowest first-guess level. The relevant code for relative
humidity read as follows -- the code for specific humidity is similar.

  if (target_z < z_fg(1,iCell) .and. k < nVertLevels) relhum(k,iCell) = relhum(k+1,iCell)

If first-guess levels are not given in bottom-to-top order, then z_fg(1,iCell)
does not necessarily contain the height of the surface in the first-guess data,
resulting in a copy of vertically interpolated relative humidity level k+1 to
level k.

One possible fix for this issue might be to compare target_z with
sorted_arr(1,1). Since sorted_arr is always sorted in ascending order by its
first index, the check to decide when to copy/extrapolate relative humidity
would be independent of the order in which first-guess levels were provided.
If the comparison were to be made against sorted_arr(1,1) rather than against
z_fg(1,iCell), the effect would be to copy the lowest *interpolated* level
downward to model levels below the first-guess ground.

An alternative fix adopted in this commit is to simply delete the logic to
explicitly copy/extrapolate downward, since both relative humidity and specific
humidity are vertically interpolated with extrap=0, which specifies constant-
value extrapolation. In this case, the lowest *first-guess* level -- rather than
the lowest *interpolated* level -- is copied downward.

Note: The issue being addressed by this commit is a dependence on the order in
which levels are given in the input intermediate file, and not on the direction
in which the first-guess model indexes its levels.