Report time to read initial state stream in atmosphere log file
This commit adds timers around the call to mpas_stream_mgr_read that is
responsible for reading the initial state for MPAS-Atmosphere (either from
the 'input' stream or from the 'restart' stream). This timing info is written
to the atmosphere log file along with several other new messages that indicate
that the initial state will be read and that the initial state has been read.
These messages may help a user to more easily identify when the model has
failed when reading ICs, or whether a model failure is happening before or
after the reading of ICS. The new log output might look like the following,
for example, when doing a restart simulation:

 Reading initial state from 'restart' stream
  Timing for read of 'restart' stream: 21.2681 s
  ----- done reading initial state -----