Switch from unencrypted Git protocol to HTTPS when obtaining MPAS-Data repo
On 15 March 2022, GitHub removed support for the unencrypted Git protocol (i.e.,
for obtaining repositories with URLs beginning with git://). As part of the
MPAS-Atmosphere build process, the checkout_data_files.sh script first tries to
obtain the MPAS-Data repository, which contains look-up tables used by physics
schemes, using the URL git://github.com/MPAS-Dev/MPAS-Data.git. Doing so results
in the messages

  *** Trying git to obtain WRF physics tables ***
  Cloning into 'MPAS-Data'...
  fatal: remote error:
    The unauthenticated git protocol on port 9418 is no longer supported.
  Please see https://github.blog/2021-09-01-improving-git-protocol-security-github/ for more information.

The simple solution is to switch to the use of HTTPS when obtaining the
MPAS-Data repository.