Add OpenACC variables in top-level Makefile to ftn, pgi, and gfortran targets
In preparation for adding OpenACC changes, this commit adds explanation
of the OPENACC=true build option in the errmsg target as well as adding
the OPENACC, CFLAGS_ACC, and FFLAGS_ACC variables to supported,
non-deprecated build targets.

When building with OPENACC=true, CFLAGS_ACC and FFLAGS_ACC are appended
to CFLAGS and FFLAGS that are passed on to the other Makefiles in src.
The _ACC variables allow addition of flags that are specific to OpenACC
compilation and are blank except for the pgi target, which inherits the
flags from the atmosphere/develop-openacc branch.