Remove unnecessary initialization of 'scalars' array in init_atm_case_gfs
Since every registry-defined field is initialized after allocation (with a
default value of 0), the initialization of the 'scalars' array in
init_atm_case_gfs is unnecessary and can be omitted. Furthermore, since the
'scalars' array is not needed for every stage in the real-data initialization
cases, future code changes could avoid allocation of the 'scalars' field
altogether in some cases, making the initialization of this field at the
beginning of the init_atm_case_gfs routine potentially problematic; removing
the initialization can therefore avoid the potential write to an unallocated
field (again, pending future code changes).