Changes for the extrapolation of temperature and horizontal velocity in real-data initializations.
(1) Reset the default for the lower air-temperature extrapolation
(config_extrap_airtemp) from "linear" to "lapse-rate" in the namelist.
(2) Set the condition for the lower extrapolation of the horizontal
velocity such that it returns the lowest analysis level value instead
of a linear extrapolation when the requested level is below the
analysis level.

Change Dynamics 1.8 for the Version 8 release.