Deallocate memory for static field tiles after they are processed
The algorithm for static field interpolation essentially performs a depth first
search for static field tiles that overlap with the region of cells assigned to
a particular MPI task. It searches for tiles, loads them, finds the closest cell
center in the kd tree, and, if there is a pixel within the cells, it adds the
tile's neighbors to the search stack.

Previous to this commit, all static tiles that overlap with the search area, as
well as their neighbors, are loaded in to memory. The memory is not deallocated
until the entire algorithm has finished searching, even though each tile is only
passed to the accumulation method once. This led to fairly enormous memory
overhead in static field interpolation, causing the init_atmosphere core to
crash under specific circumstances.

This change deallocates the tile memory directly after marking the tile as
processed. This significantly reduces the memory overhead, as only unprocessed
tiles remain in memory. In the future, changes may be added to allow
geotile_manager to load tile data only on request, reducing memory usage even
further. However, this commit is a simple, practical, and direct solution.