Refactor the setup and use of group halos in the atmosphere core
Prior to this commit, halo exchange groups in the atmosphere core were set up in
the atm_core_init and mpas_atm_dynamics_init routines, some were destroyed later
in the atm_core_init routine, and others were destroyed in the atm_core_finalize
and mpas_atm_dynamics_finalize routines. When halo exchange groups were
communicated, different parts of the solver (initialization, dynamics, physics)
made direct calls to mpas_dmpar_exch_group_full_halo_exch.

In an effort to provide a single place where halo exchange groups are set up and
destroyed, and also to abstract the details of which specific routine is used to
exchange a halo group, this commit:

  1) Moves the setup of all halo exchange groups to a new routine,
     atm_build_halo_groups, that is called from atm_core_init;

  2) moves the destruction of all halo exchange groups to a new routine,
     atm_destroy_halo_groups, that is called from atm_core_finalize;

  3) defines an abstract interface for routines that can exchange halo groups;

  4) uses a procedure conforming to this abstract interface to exchange halo
     groups during initialization, dynamics, and the transfer of physics
     tendencies to the dynamics.

The specific routine to exchange a halo group is set in the
atm_build_halo_groups routine as a procedure pointer in the atm_core module, and
this routine is passed as a procedure argument to the dynamics and thence to the
physics_get_tend routine.

The end goal of the refactoring in this commit is to simplify the work needed to
use alternative halo exchange methods in the atmosphere core.