Add new namelist option, config_halo_exch_method, to select halo exchange method
In preparation for the introduction of alternative methods for exchanging halos,
this commit introduces a new namelist option, config_halo_exch_method, that can
be used at runtime to choose the halo exchange method.

The new config_halo_exch_method resides in a new namelist group, &development,
that is hidden by default. At present, the default and only value for this
option is 'mpas_dmpar', which selects the group halo exchange methods from the
mpas_dmpar module.