Add nominalMinDc to init_atmosphere core, and scale its value for each init case
The init_atmosphere core now includes a new scalar variable, nominalMinDc, which
specifies the nominal grid distance in the highest resolution part of the mesh
(where mesh_density == 1.0). The nominalMinDc variable is read from the "input"
stream and written to the "output" stream. The value of this variable is also
scaled by each each initialization case: for real-data initialization the value
is scaled by the Earth's radius, 6371.229 km, and for idealize cases the value
is scaled by a factor that depends on the specific case.