Add mpas_modify_att() subroutines to mpas_attlist
CF compliance requires certain attributes to be updated periodically throughout
a run. The current attlist module only supports creation and retrieval of
attributes. This commit addds this functionality, and some basic tests. The
subroutines work very similarly to the "mpas_get_att" subroutines, but modifies
the contents of the linked list node if the type and name match the expected
type and name. If they do not, the subroutines will return a 1 in ierr.

It is also noted that these procedures are not currently threadsafe. Future
work may include threadsafe attribute modification and a more comprehensive
test suite for all of the attlist procedures.