Add missing physics_mmm module search path in atmosphere Makefiles
Some compilers (e.g., nvfortran) require the paths of modules that are used
indirectly to be present in the module search path provided with the -I...
option.

This commit adds -I./physics/physics_mmm to the set of compiler flags in
src/core_atmosphere/Makefile and -I../physics/physics_mmm to the set of compiler
flags in src/core_atmosphere/dynamics/Makefile to fix compilation errors with
certain Fortran compilers.