Fix out of bounds access in threadErrs on failure
The "threadErrs" array is intended to hold the error code for each thread
running a bank of tests. In the field tests, the thread number currently
running the function is retrieved by "mpas_threading_get_thread_num", and
stored in "threadNum". If the threadnum is 0, the tests are actually executed.
However, because arrays in fortran are 1-indexed, when the tests fail and
attempt to set threadErrs( threadNum ) to 1 it attempots to access
threadErrs( 0 ), which is out of the array bounds. This is fixed by adding 1 to
the thread number- this is done elsewhere in the halo exchange tests, but was
omitted in this test suite.