mpas_split_string now trims strings before splitting
This commit adds a step to the mpas_split_string subroutine. In the process of
adding tests, it was discovered that splitting a string while using a space as
a delimiter causes the string to be split on every trailing space, leading to
a splitStrings array that is equal to the size of the untrimmed input string,
minus the number of non-space characters, plus one. This behavior is unwanted.
This has been solved by trimming off trailing spaces before splitting the
string.