Add cf-compliant time axis to core_atmopshere
CF compliance in time axis requires a unit of '<time unit> since YYYY-MM-DD
hh:mm:ss'. This commit creates the 'Time' variable in the core_atmosphere
Registry.xml and changes the 'units' attribute to 'seconds since
<start_timestamp>'. The start timestamp is modified from the initial time format
to remove the underscore and replace it with a space to match the CF compliant
time string format.

Time variable is then updated at the same time as the xtime variable by adding
dt' to the Time at time level 1, and saving it to time level 2. This produces a
CF-compliant time variable that tracks xtime at each step.