Add cf-compliant "Time" variable to core_init_atmosphere
This commit adds the "Time" variable to core_init_atmosphere Registry.xml for
LBC's, surface updates, and initial conditions. This is the same Time variable
that is used in core_atmosphere as a cf-compliant time variable. The "units"
attribute is also updated to match the initial_time of the initial conditions.

This commit also adds cf-compliant time to the outputs for both lbc files and sfc
update files. This time is derived by getting the time interval between
MPAS_START_TIME and the current time, and writing this interval in seconds to
the Time variable for both the lbc files and the sfc update files.