Properly set return error codes and exit from some mpas_io routines with SMIOL
In a subset of the routines in the mpas_io module, return error codes were not
properly set when errors occurred in calls to the SMIOL library, and those
routines didn't exit at the point of those unrecoverable errors.

This commit updates various mpas_io routines so that they set the 'ierr' return
code and return when MPAS is using the SMIOL library.