Add new 'ioContext' argument to MPAS_io_err_mesg to provide additional context
The messages that are printed for some errors by the MPAS_io_err_mesg() routine
may benefit from having access to the I/O context within which the error
occurred.

For example, the SMIOL library can use the SMIOL context member of the MPAS
'mpas_io_context_type' type to provide messages from the specific library that
was called by SMIOL. Now, if the SMIOL error code is SMIOL_LIBRARY_ERROR, an
additional line is printed by the MPAS_io_err_mesg() routine; for example:

  ERROR: Library error message: Unknown error in file operation

With the addition of the new ioContext argument to MPAS_io_err_mesg, calls to
this routine from elsewhere in MPAS (currently only in the mpas_io_stream
module) have been updated to provide an appropriate value for this new argument.