After reading a stream, only reindex fields that were active in that stream
Historically, input streams that contained mesh indexing fields (e.g.,
cellsOnCell) did not attach packages to those fields in the stream, and so we
could safely assume that all indexing fields in an input stream should be
converted to local indices via a call to the postread_reindex routine after the
stream was read.

In future, however, packages may be used to conditionally read mesh indexing
fields from a stream, in which case it's important to only reindex those fields
that were actually read; otherwise, indexing fields that already contained
correct local indices will be corrupted.

With this commit, two new pools -- the allPackages pool and the field_pkg_pool
-- are passed to the postread_reindex routine, which can then determine whether
each indexing field in an input stream was active and therefore needs to be
reindexed.