Add in_defaults=true to invocation of streams_gen for the atmosphere core
When running the streams_gen program for the atmosphere core (from
src/core_atmosphere/Makefile), the command-line argument in_defaults=true is
provided, so that any stream that has the "in_defaults" attribute will only be
written to the default streams.atmosphere file if the value of that attribute is
"true". Note that for any streams that do not specify the "in_defaults"
attribute, those streams will always be written to the streams.atmosphere file.