Fix nvfortran internal compiler error for streamInfo % finalize in mpas_finalize
The nvfortran 23.1 and 23.5 compilers fail with an internal compiler error in
the mpas_finalize() routine in the call to domain_ptr % streamInfo % finalize():

NVFORTRAN-S-0000-Internal compiler error. check_member: member arrived with wrong derived type    5622  (mpas_subdriver.F: 443)

(Note that what was line 443 is now line 446.)

The simple work-around in this commit is to declare a local pointer to
domain_ptr % streamInfo, and to then call the finalize() routine from that local
pointer.