Split physics source compilation from library updates for parallel builds
This commit splits the compilation of source files in MPAS-A physics from the
updating of libphys.a to support parallel builds of MPAS-A. Previously,
compilation of source files in physics_wrf, physics_noahmp, and physics_mmm
included updates of libphys.a with the resulting object files. When physics_wrf,
physics_noahmp, and physics_mmm are compiled in parallel, this could result in
race conditions when updating libphys.a.

Now, Makefiles have been modified so that source files in physics_wrf,
physics_noahmp, and physics_mmm can be compiled in parallel first, and as a
subsequent step, object files from these directories are added to libphys.a in
a serial fashion.

Note that since the physics_mmm directory is populated from the external
MMM-physics repository, this commit updates the MMM-physics tag to
20240626-MPASv8.2, which includes updates to the MMM-physics Makefile.mpas.